Fixnetwork

+++ src/ScenarioTest.ResourceManager/EnvironmentSetupHelper.cs
@@ -672,7 +672,7 @@ namespace Microsoft.WindowsAzure.Commands.ScenarioTest
 #if NETSTANDARD
                     var moduleShortName = moduleName.Split(new string[] { "\\" }, StringSplitOptions.None).Last().Split(new string[] { "/" }, StringSplitOptions.None).Last();
                     powershell.AddScript("Enable-AzureRmAlias -Module " + moduleShortName.Substring(0, moduleShortName.Length - 5));
-                    if (moduleShortName.Equals("Az.Storage.psd1"))
+                    if (moduleShortName.Equals("Az.Storage.psd1") && modules.Any(s => s.EndsWith("AzureRM.Storage.ps1")))
                     {
                         powershell.AddScript("Get-Alias | Where-Object {$_.Name -like '*-AzureRmStorage*'} | ForEach-Object { Remove-Item \"alias:\\$_\" }");
                     }
